% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_opener.R
\name{oeis_browse}
\alias{oeis_browse}
\title{Open oeis web page on a given sequence}
\usage{
oeis_browse(x)
}
\arguments{
\item{x}{a string with the name of the sequence
in the oeis data base
(which can be fond using the oeis_finder() function).}
}
\value{
the webpage is opened on the default browser
}
\description{
Open oeis web page on a given sequence
}
\details{
It opens the oeis web page on the sequence that matches
the name provided by the user.
}
\examples{
oeis_browse("A000045")
}
