% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadAndCache.R
\name{oeis_cache}
\alias{oeis_cache}
\title{Downloads and caches OEIS data}
\usage{
oeis_cache()
}
\value{
A message indicating how to view the downloaded data
}
\description{
Downloads and caches OEIS data
}
\details{
When called, this function downloads the oeis data directly from the oeis
website and saves it in the cache memory of the user's computer for easy
and rapid access.
}
\examples{
oeis_cache()
}
