% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis_download_webpage.R
\name{oeis_download_webpage}
\alias{oeis_download_webpage}
\title{Download and parses the OEIS webpage of a sequence}
\usage{
oeis_download_webpage(
  x,
  portion = c("all", "references", "links", "comments", "formula", "examples",
    "second", "third", "fourth")
)
}
\arguments{
\item{x}{This argument is the id for a mathematical sequence}

\item{portion}{This argument specifies what data the user wants to see. This could be just the references, links, or the entire webpage. Alternatively, users can select second, third, or fourth to download those respective sections. 1st is generally offset, and second refers to the second section header in the information section of a sequence entry.}
}
\value{
All of the text such as references, various links, and comments is returned depending on what the user selects.
}
\description{
Download and parses the OEIS webpage of a sequence
}
\details{
This function uses the rvest package to download references, various links, and comments from the webpage of a sequence on the OEIS website. If the OEIS sequence page does not follow a standard structure, Users can access the second, third, or fourth sections which include important information about the sequence.
}
\examples{
oeis_download_webpage("A000045", "all")
}
