% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis_finder.R
\name{oeis_finder}
\alias{oeis_finder}
\title{Searches for numeric sequences}
\usage{
oeis_finder(x)
}
\arguments{
\item{x}{a numeric vector with the first few numbers of the desired sequence}
}
\value{
a tibble with all the sequences that begin with the inputted numbers,
and their corresponding name in the oeis data base
}
\description{
Searches for numeric sequences
}
\details{
Searches through the oeis data base for sequences
that match the first few numbers in a sequence (inputted by the user)
}
\examples{
oeis_finder(c(0,1,1,2,3,5,8,13,21))
}
