% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadAndCache.R
\name{oeis_load}
\alias{oeis_load}
\title{Displays oeis data}
\usage{
oeis_load()
}
\value{
the oeis data
}
\description{
Displays oeis data
}
\details{
Allows the user to access the oeis data saved in the cache memory.
If the data has not been downloaded, it tells the user he/she has to
run the oeis_cache()function first.
}
\examples{
oeis_load()
}
