% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep.R
\name{oeis_play}
\alias{oeis_play}
\title{Play music from the OEIS}
\usage{
oeis_play(x, low = 100, high = 4000, wav = NULL)
}
\arguments{
\item{x}{The OEIS sequence name.}

\item{low}{Lowest frequency.}

\item{high}{Highest frequency.}

\item{wav}{The default play for Wave files. If you are using mac, then
this is likely "afplay" or the location of "afplay". If you are
using Windows, it's probably OK to leave this as NULL. If you are using
Ubuntu, I got this to work using "aplay".}
}
\description{
Uses the tuneR package to encode each integer as a note. Solution
based on this StackOverflow: \url{https://stackoverflow.com/questions/32857065/is-it-possible-to-code-music-in-r-and-play-it-back-mac-os-x}
}
\examples{
\dontrun{
  ## Using wav = NULL works on Windows:
  oeis_play("A108618")
  oeis_play("A007318")
  oeis_play("A025480")
  oeis_play("A056239")
  oeis_play("A117153")
  oeis_play("A000005")
  oeis_play("A003602")
  oeis_play("A004718")
  oeis_play("A117154")
  oeis_play("A000010")
  oeis_play("A064413")
  oeis_play("A006577")
  oeis_play("A005132")
  oeis_play("A001223")
  oeis_play("A119953")
  oeis_play("A123456")
  
  ## On Ubuntu, I got it to play music using aplay
  oeis_play("A123456", wav = "aplay")
  oeis_play("A117153", wav = "aplay")
}
}
\author{
David Gerard
}
