% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{oeis_plot}
\alias{oeis_plot}
\title{Plots oeis numeric sequence}
\usage{
oeis_plot(x, n = NULL, title = x)
}
\arguments{
\item{x}{a string with the name of the desired sequence
on the oeis data base}

\item{n}{an integer indicating the number of elements of
the sequence that should be plotted.
If null, then all the complete oeis sequences is plotted}

\item{title}{a string indicating what the title of the plot shoudl be}
}
\value{
a line plot displaying the sequence
against the corresponding indices
}
\description{
Plots oeis numeric sequence
}
\details{
Plots the oeis sequence that matches the name given by the user
}
\examples{
oeis_plot("A005132")
}
