% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{oeis_plot2}
\alias{oeis_plot2}
\title{Plot two oeis numeric sequences}
\usage{
oeis_plot2(A, B, title = paste(A, "vs", B))
}
\arguments{
\item{A}{a string with the name of the first desired sequence
on the oeis data base}

\item{B}{a string with the name of the second desired sequence
on the oeis data base}

\item{title}{a string indicating what the title of the plot shoudl be}
}
\value{
a line plot displaying the two sequences
against the corresponding indices
}
\description{
Plot two oeis numeric sequences
}
\details{
Plots the oeis sequences that match the names provided by the user
}
\examples{
oeis_plot2("A297346","A297347")
}
