% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis_retrieve.R
\name{oeis_retrieve}
\alias{oeis_retrieve}
\title{Retrieve oeis sequence}
\usage{
oeis_retrieve(x)
}
\arguments{
\item{x}{a string with the name of the desired sequence
in the oeis data base}
}
\value{
a numeric vector with the elements of
the desired sequence
}
\description{
Retrieve oeis sequence
}
\details{
Searches through the oeis data base for a sequence that
matches the name inputted by the user
and returns it as a numeric vector
}
\examples{
oeis_retrieve("A000045")
}
